fix: foreign alloc return values
When a return value is e.g. :string, the memory for the string is allocated and owned by SBCL. Thus
the caller is unable to free it. The memory will eventually be freed by GC, which is not good as
the caller could use it after it is GC'd. This change makes :string be foreign-allocated and thus
can be freed with `free`.